+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 0.3.0 2024-01-21
+
+- Breaking: Removes async-trait dependency
+- Breaking: Redis backend now uses BITFIELD to store counts
+
 ## 0.2.2 2022-04-19
 
 - Improve documentation.